+++ packages/exchange/package.json
@@ -24,7 +24,16 @@
   ],
   "type": "module",
   "types": "dist/index.d.ts",
-  "exports": "./dist/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./schemas": {
+      "types": "./dist/schemas.d.ts",
+      "default": "./dist/schemas.js"
+    }
+  },
   "publishConfig": {
     "access": "public"
   },
