+++ packages/exchange/src/schemas.ts
@@ -1,0 +1,51 @@
+import {z} from 'zod';
+
+/*
+ * Browser-safe schemas, exposed as the `@typedtrader/exchange/schemas` subpath. This module must
+ * not import broker code (or anything else touching Node.js built-ins), so bundlers can ship
+ * candle validation to the browser without dragging the whole exchange package along.
+ */
+
+export const CandleBaseSchema = z.object({
+  /** ID of base asset */
+  base: z.string(),
+
+  /** ID of quote asset */
+  counter: z.string(),
+
+  /** True, if this candle is the latest / current candle from the exchange. This flag is good to know if a candle comes from a history import or not. */
+  isLatest: z.boolean().optional(),
+  /** Bucket start time in simplified extended ISO 8601 format */
+  openTimeInISO: z.string(),
+  /** Bucket start time converted to milliseconds (note: Coinbase Pro actually uses seconds) */
+  openTimeInMillis: z.number(),
+  /** Candle size in milliseconds */
+  sizeInMillis: z.number(),
+});
+
+export type CandleBase = z.infer<typeof CandleBaseSchema>;
+
+export const CandleSchema = z
+  .object({
+    /** Closing price (last trade) during the candle interval */
+    close: z.string(),
+    /** Represents the minimum price that a seller was willing to take at the end of the candle */
+    closeAsk: z.string().optional(),
+    /** Highest price during the candle interval */
+    high: z.string(),
+    /** Represents the minimum price that a seller was willing to take at the peak of the candle */
+    highAsk: z.string().optional(),
+    /** Lowest price during the candle interval */
+    low: z.string(),
+    /** Represents the minimum price that a seller was willing to take at the bottom of the candle */
+    lowAsk: z.string().optional(),
+    /** Opening price (first trade) during the candle interval */
+    open: z.string(),
+    /** Represents the minimum price that a seller was willing to take at the beginning of the candle */
+    openAsk: z.string().optional(),
+    /** Amount of traded base currency during the candle interval */
+    volume: z.string(),
+  })
+  .merge(CandleBaseSchema);
+
+export type Candle = z.infer<typeof CandleSchema>;
